change Location float32 to float64

+++ gateway/api/types.go
@@ -2,6 +2,6 @@ package api
 
 type Location struct {
 	ID   string  `json:"id"`
-	Lat  float32 `json:"latitude"`
-	Long float32 `json:"longitude"`
+	Lat  float64 `json:"latitude"`
+	Long float64 `json:"longitude"`
 }
